"PEOPLE YOU MAY KNOW"
Adding algorithm for "PEOPLE YOU MAY KNOW"  for a larger dataset. It suggest people to a person based on connections with friends and mutual friends.New users are suggested based on Priority list i.e the more the number of mutual friends,the more chances of coming it at top.